moved calculation classes to Util folder
