+++ GameOfLife.Data/Util/GameCreationService.cs
@@ -1,12 +1,12 @@
 ﻿using GameOfLife.Data.Constants;
 using GameOfLife.Data.Entities;
-using GameOfLife.Data.Entities.Game;
+using GameOfLife.Data.Entities.Games;
 using GameOfLife.Data.Entities.Rules;
 using GameOfLife.Data.Interfaces;
 using GameOfLife.Data.Interfaces.Game;
 using GameOfLife.Data.Interfaces.UI;
 
-namespace GameOfLife.Data.Services
+namespace GameOfLife.Data.Util
 {
     public class GameCreationService(IInputHandler inputHandler) : IGameReceiver
     {
